Delegate extension of worklist to modifyA implementation
I found a fix for the symbolic pruning that doesn't require that every
new node is added to the worklist: the modifyA function must add the
parents of the pruned class to the worklist for them to be upward
merged. I think it's a good compromise for requiring the user to do
this. Adding the added node to the worklist everytime creates too much
unnecessary work.