+perf: Fix deduplication, use list for worklist and related improvements
We were previously using a Map for the worklist, and the deduplication
step wasn't being done correctly. This came to be from historic changes
to the representations that ultimately led to deduplication being
overlooked.

The performance significantly improves for some tests with this change.

Improve the e-class parents representation with a sized list (which is
like a list, but with O(1) size function)

De-duplication results: it seems to be actually slower to `nubOrd` than not

Other related bugs fixed