Improve support for monadic Analysis
In a previous commit we introduced functions to facilitate usage of
monadic analysis by setting the domain to (m a) and traversing over m.

Unfortunately, in practice, setting the domain to (m a) requires an
instance of Eq (m a) to define the analysis, which is much harder for a
monadic context.

We improve support for monadic Analysis by providing a new type-class
'Data.Graph.Analysis.Monadic' which defines its operations over a
monadic context, while keeping the domain non-monadic, and by changing
the operationM variants of operations to require a Monadic Analysis
instead of a standard Analysis.