Fix PatId collisions and clean up Database slightly
Make VariablePatterns literal Strings.
Now, preserve a mapping from Strings to 'Var' and get that result out of
compiling the query (slightly earlier) to pass to the remaining
operations which need to de-reference user-written names.

Add a newtype for 'Var' and a newtype for 'Subst' to make sure we don't
conflate ClassIds with Vars. It looks like there were no mistakes here,
but the code is cleaner for it.

Fixes #20 and #32