Saturation: Ensure that banned rules are retried on saturation
Previously when we would reach saturation we would fail to consider
whether previously-banned rules might now be applicable. This led to
missed rewrite opportunities.

The solution taken here is simple: reset the rule stats to remove all
bans and try one final iteration. Another approach that I considered is
to instead introduce a way of notifying the scheduler that we have
reached a fixed-point, making it responsible for unbanned. However, it's
hard to see what the scheduler would do beyond what is done here.

Closes #45.